Canvas order changed, mouse events
The second canvas (the feedback canvas) is now the ANNOTATION CANVAS. It
holds all the drawn objects.

The top canvas is still the interaction canvas. The reason for this is
so that the edit canvas idea and the feedback canvas can now be
consolidated into ONE canvas, furthermore, the draw for the indicators
will now appear above the previously drawn objects, which looks nicer
and makes more sense.

The mouse events are now bound to the document, and only when the mouse
is on the canvas will any canvas events actually occur, per a check on
mouse boundaries on the document vs the canvas position. This works
better since now all three canvases can listen to one event rather than
having the first canvas trigger the other canvases (which essentially
requires two events).