Basic Check, Parser and Storage skeletons
Created a function (basicCheck) to check the objects of canvas fed to
it, then sends them to a storage handler to be stored. I also made a
skeleton storageHandler for storing and a skeleton Parser for
identifying the type of canvas in order to understand what attributes
need to be stored.