Added anchor edit function
In edit mode, clicking on an anchor after selecting a shape will now
allow the user to edit and save the changes made to a single anchor.
There is feedback for what the new shape will look like. Currently there
is a bug where it will always trigger after the second click if it is in
the snap zone, but this can be fixed easily.