Minor bug fixes to anchor edit
Made some small logic changes so that editing the entire shape and
editing an anchor is now more explicit.

Still has the bug where the next click on an anchor switches modes: it
will redraw the entire shape but only update the anchor coordinates.

Also, because closing a path adds an anchor that is equal to the first
anchor, it is possible to open the shape up by clicking on the last
anchor and trying to move it.