Fixed bugs with editing
Trying to change the anchor on a shape now only fires after the shape is
selected but before the shape has been moved.

Moving the anchor works and saves correctly. Had to remove the last
duplicate point (used for SVG tags) to edit correctly so the user could
not open the shape up - this may need to be changed. The point is added
back into the completedPaths upon save.

Hitting space will switch shapes if they overlap and if no shape or
anchor has been moved on the currently selected shape. It still selects
the smallest shapes. The page still moves, however.