Major changes: included parser
Moved the JSONparser tool into the main program here. Text area on
proto.html now sets up to read in a link or json and parse it as best it
can.

CHandler now tries to find the associated data it needs from the canvas
json and load it in the CONFIGS object, and update the display
accordingly. If it cannot find the canvas dimensions  (for instance, an
annotationList or annotation id were given), it will reattempt to find
the data associated. For now, it falls back to resizing the canvas to
1000 x 1000 by itself.

Currently draws no annotations, but does find all associated annotations
in rectalinear form and prints out their dimensions in the console. We
will be able to use this to draw annotations on the screen shortly.