Scans and draws for rectalinear annotations from parsed data
The program receieves the data from the parser and then finds which ones
are SVGs or rectalinears. SVGs detect some basic supported types but
none are currently drawn.

Rectalinear annotations, however, are drawn, even if the canvas
dimensions could not be found (ie only an annotationList was entered).

Still need to convert the coordinates on saving the object if the
dimensions do not match the display resolution of the canvas.

Cannot currently edit the drawn annotations - this is because they are
not enumerated in the completedPaths list, which will come later when we
extract and save the data to this list after drawing the annotations.
Redraws will use the current method after parsing the data for each
annotation, which uses the anchorList data structure to draw the shapes
for interaction.