Setup ajax, new dummy canvas, parsing triggers
Setup the ajax button on the toolbar only for dummy canvases.

Load Test Canvas Handler now loads a dummy canvas with some dummy
annotation data in it. You can draw to this canvas and click export
(currently doesnt work because of cross domain issues)

Parser now lets the handler know when the parsing is finally complete,
so that the handler can now begin loading data or handling errors