Canvas and anno resize implementation
Canvas now targets to fit between some min and max dimensions.

Also resizes any incoming annotations based on the scale factor that the
canvas was resized by.

No rounding occurs until the JSON is generated per annotation to reduce
loss of precision but to also make sure that the data conforms to IIIF
and OA.

Annotation viewer displays the native points (that is, the points that
conform to original canvas dimensions) but do NOT reflect the points
that are shown on screen, which are pre-adjusted.