Added Line Color Change Function
I have added functions to change the color of the line via buttons on
the tool bar. Each button is linked to a separate function in the tool
bar that calls a single function in the canvas handler that changes the
color. At present, only one button has been placed in the toolbar, but
it does not change the color.