Parser fixes, SVG conversion fixes, attempt better export
Changed some indicators in the parser. Now checks to see if the url
associated with an annotationList was resolved (or has annotations in
embedded within the original JSON for the canvas), and adds it to the
list of annoLists (in case there are multiple). Also added a trigger in
basicCheck to tell the CH if it found a key "@type" with the
"sc:AnnotationList" value, to indicate we have annotations (this is in
case we dont have to resolve a URI to retrieve the annotations).

SVGToAnchor function now works properly

Export function is now setup to handle multiple annotationLists, and the
anchorList now keeps track of the index where the anno was located in
this list, as well as which annoList it belongs to (by index). Export is
not complete and not tested yet.