more json display fixes, minor conversion fix
better logic for JSON display, that now seems much more stable and does
not duplicate items on the json display or cause weird event errors

createSVGTag also changed the anchors to the native resolution, which is
drastic if the resolution of the canvas was adjusted, since the anchors
were passed by reference. fixed by cloning the structure before
converting.