Save Indicator Proper explanation
I accidentally pushed before giving a full explanation. A variable in
canvasHandler changes a bool which keeps track whether or not there are
unsaved changes, and compares this value to the indicator in
canvasHandlerToolbar. If they differ, the indicator changes. The only
problem currently is that the previous indicator is not deleted, this
will be fixed shortly.