save modal, annotation text styles
a save modal now pops up and disables any user interaction until the
save is complete, whereupon the modal is hidden again.

The text areas for editing annotation text now sizes a little more
properly, and the toolbar has a minimum width until it jumps to the end
of the page.