Fixed delete issues again
On delete, we only checked if the completedPath had an annoId to splice
it from the list - this was changed so that when call to remove marked
for deletion, the splice always happens.

We also waited until all the annoItemList items for the toolbar were
updated before removing them from the jsonDisplay and also from
annoItemList. before we were splicing from the list during the loop,
causing duplication errors.